Make folly::Uri::port() return uint16_t instead of uint32_t.
Summary: We can safely assume that ports are always going to be 16bit (TCP or UDP), so no need to have a uint32_t here.

Test Plan: Ran tests in folly/tests.

Reviewed By: tudorb@fb.com

FB internal diff: D973291

+++ folly/Uri.cpp
@@ -79,7 +79,7 @@ Uri::Uri(StringPiece str) : port_(0) {
 
     StringPiece port(authorityMatch[4].first, authorityMatch[4].second);
     if (!port.empty()) {
-      port_ = to<uint32_t>(port);
+      port_ = to<uint16_t>(port);
     }
 
     username_ = submatch(authorityMatch, 1);
